Bug 1885619 - Add --power-test flag to raptor for mobile power testing. r=perftest-reviewers,aglavic
This patch adds a `--power-test` flag that can be used to do power usage testing on mobile devices that are hooked up to a USB power meter. It is gathered for all tests when it's enabled, and a helper method is added to make it easier for support-scripts to gather it. An update is also done to browsertime to gather the power usage data. At the same time, a bit of documentation about this flag is added to our Raptor Browsertime docs.

Differential Revision: https://phabricator.services.mozilla.com/D212107

UltraBlame original commit: 480fd2f7a71b039bb6266915087d74d3fd89f7b0