Bug 1898621: Provide error results for surface promotion methods. r=gw,nical
This uses the warn macro for promotion errors. It's not clear this will
be useful for web authors, but hopefully it's useful for Firefox
developers.

This has the potential to be very spammy. I need ideas on how to limit
the reporting to new or interesting errors.

Differential Revision: https://phabricator.services.mozilla.com/D212340

UltraBlame original commit: a06f8745fb4ec5ed8e28c7704f21c9b94bf1e21f