Bug 1901257 - Add a visualization for last 20 slow scrollign frames. r=gfx-reviewers,gw
This patch adds the "Slow scroll frames" query in the profiler HUD, relying on the existing ProfilerFrameCollection code.
It also changes the visualization to sort labels alphabetically so that the breakdown of the slow frames can be listed in the order of the events in the frame.
Frame building-related events are in blue-ish tones while renderer ones are in red-ish tones.

It also changes the classification of scrolling frames to consider frames that do not have scene building instead of looking for frames that have animations.

Differential Revision: https://phabricator.services.mozilla.com/D212936

UltraBlame original commit: dc340c7370a6b36569f9c2099791cfa6c5636a09