Bug 1899339 - Part 1: Ignore cross zone edges that are already sufficiently marked r=sfink
This assertion was complaining about trying to mark the source of a cross-zone
edge to a nursery thing gray when in fact the source was already marked black.

The patch checks whether the target is already sufficently marked and bails out
early in that case. This can happen due to a barrier marking a thing black when
it's already on the gray marking stack.

Differential Revision: https://phabricator.services.mozilla.com/D212538

UltraBlame original commit: ca64977976882320d4d06c4603bfe974756d3c2d