Bug 1844264 - [devtools] Make shape editor toggle button keyboard accessible. r=devtools-reviewers,ochameau.
Turn the icon into a proper button, add a title to it
and properly manage its aria-pressed attribute.
We no longer set the active class, and instead rely
on the aria-pressed attribute in the CSS and in tests.

Differential Revision: https://phabricator.services.mozilla.com/D213032

UltraBlame original commit: 88308f895522ad817a5cda5f27e9f66cea2aee71