Bug 1893527 - Change menu button visibility based on that of navbar r=android-reviewers,petru
Fixed showing menu for home page as well, as the whole state of the toolbar depended on one flag. Previously, it didn't check if the device is in landscape mode or a tablet, just relying on the feature flag. Now it considers all the conditions to set the toolba state on the homepage.

Differential Revision: https://phabricator.services.mozilla.com/D211639

UltraBlame original commit: 9c94f26161d8de3f1e304a3ccb3a23b8ff489016