Bug 1900222 - Add test coverage for content scripts in slow-loading frames r=zombie
The reported bug is caused by an attempt to schedule a content script in
an initial about:blank document/frame for which document_end and
document_idle never resolve. Although the bug is about dynamic execution
(insertCSS/executeScript/etc), the underlying injection logic is also
shared by declarative content scripts.

This test adds test coverage for two relevant scenarios:

- Content scripts declared via content_scripts in manifest.json (this is
  also the underlying implementation of contentScripts.register,
  userScripts.register and scripting.registerContentScripts).
  This WORKS AS EXPECTED, because scripts never run in initial
  about:blank ( which is reported at bug 1415539 + bug 1486036 ).

- Content scripts executed in existing content as part of extension
  install via ExtensionPolicyService::InjectContentScripts. This
  UNEXPECTEDLY results in executions that are stuck and never-settling.

The same issue is also relevant to the handleActivateScripts logic,
which did not and still does not have explicit coverage for this
scenario. A generic solution to the InjectContentScripts case will also
apply to handleActivateScripts, so I did not add tests for that here.

A test case for the dynamic code execution case is not part of this
patch because the observed result would be that the test gets stuck.
This test case will be included with the next patch that fixes this bug,
which will also update the test expectation for the "UNEXPECTEDLY"
never-settling promise mentioned above.

Differential Revision: https://phabricator.services.mozilla.com/D212986

UltraBlame original commit: c018b69d1062cf1964b242718220bc08e2c95d65