Bug 1901894 - Stop scheduling content script injection for initial about:blank of top-level opaque origins r=rpl
The primary change of this patch is to add IsInitialDocument() to the
"is this an about blank document where I should be able to run scripts
despite it not having a matching origin" logic. This is to avoid
scheduling content scripts in top-level about:blank in the parent
process as mentioned in the bug report.

The original use case (supporting execution in top-level about:blank)
continues to be supported, which is demonstrated by the still passing
tests in test_ext_contentscript_about_blank.html. A new regression test
was added (test_toplevel_aboutblank_at_document_start) in the same file,
which fails before the fix  when run with --tag=in-process-webextensions
due to unexpected extra messages like the bug, and passes with the fix.

Differential Revision: https://phabricator.services.mozilla.com/D213439

UltraBlame original commit: 08e0af3a2bb15857b80f4b407099c9f3500e1a70