Bug 1896948 - restrict MFCDM support based on origin. r=media-playback-reviewers,padenot
We wanted to have a temporary workaround to explicitly allow/block
domains to use MFCDM due to the bug 1901334. As this is only a
workaround, we'd like to do it in a simplest way in order to save our
resource.

Our strategy is to keep Nightly unfiltered so that we could discover
more websites affected by the CSS problem. On other channels, we would
currently only allow Netflix to use MFCDM as we have been testing
with them and know that issue won't affect them.

We will first try to use the allow list first to ensure the MFCDM
playback quality, and gradually switch to the block list to enable the
MFCDM ability to more websites.

Differential Revision: https://phabricator.services.mozilla.com/D213820

UltraBlame original commit: f34a76aaa3c6884a1f65d492d71969c89ee55484