Bug 1902983 - Don't use bailout data after iterating Wasm frames. r=iain
This is similar to bug 1900523, but the fix there was incomplete because the
`JSJitToWasm` frame type is only used when we go through the Wasm JIT entry
trampoline. Ion can also call Wasm functions directly and in that case the type
will be `FrameType::Exit`.

Differential Revision: https://phabricator.services.mozilla.com/D214098

UltraBlame original commit: 6ef4b7955068537bda9d211548128e5e84836cfe