Bug 1904049 - Skip printing related test for shadow-crossing selection on Beta r=dholbert
We need to skip them when shadowDOM selection is not enabled because
flipping this pref with the selection around can cause unexpected
behaviours. I don't see there's a way to clear the selection
when the test is done, so let's just skip them on Beta.

Differential Revision: https://phabricator.services.mozilla.com/D214619

UltraBlame original commit: dfa3ab731b9eac2d8a6229583792a886782acfd0