Bug 1903725 [wpt PR 46837] - Improve wavy decorations for smallish fonts, a=testonly
Automatic update from web-platform-tests
Improve wavy decorations for smallish fonts

The wavy text decoration for 10pt font is not as small as one would
expect. The issue is that we cap the smallest wavy size we are willing
to draw, and this is too pessimistic.

To address this, split the limits for the height and width of the
wave and reduce them. In the process remove a function.

Bug: 347595640
Change-Id: Ibd2d2f4b37f1454a5697d38703b47bd9d5de6454
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5631324
Reviewed-by: Fredrik Söderquist <fsopera.com>
Commit-Queue: Stephen Chenney <schenneychromium.org>
Cr-Commit-Position: refs/heads/main{#1317302}

--

wpt-commits: 2f564c5b48de650aeb3daaca7688d53fdd178433
wpt-pr: 46837

UltraBlame original commit: 29d4017625dea9702e89f72664a2da0b09b4400c