Bug 1901764 - Make select label not react to events. r=dholbert
This is the same fix as bug 1893177, but for the <label>. It should
ideally not be necessary at all, but it's unclear to me why the target
difference makes us not see the pointer-up event.

That's worth digging into, and will file a follow-up on it, but this
restores the previous behavior exactly, so should be worth doing in the
interim while we figure that out.

Differential Revision: https://phabricator.services.mozilla.com/D214709

UltraBlame original commit: 556a2c727e87788703fffec2c182f6ab95621881