Bug 1903502 - Use in-tree robustcheckout.py in tasks running on generic-worker. r=releng-reviewers,sheehan,taskgraph-reviewers,jcristau
The robustcheckout.py that is in-tree ends up in docker images, used by
tasks running on docker-worker, but for tasks running on generic-worker,
we currently rely on whatever version of robustcheckout.py is there.
It currently seems out of date, but whenever we make changes to the
in-tree version, we don't end up with the fix on these jobs without some
additional manual work.

With this change, we do the same as with run-task and fetch-content: we
pull the in-tree version that we store as an artifact of the decision
task.

Differential Revision: https://phabricator.services.mozilla.com/D214234

UltraBlame original commit: 2147e0051e231c4a0e6081bf3fd72e775290a607