Bug 1841818 - Enable MOZ_LOG logging on Android. r=android-reviewers,geckoview-reviewers,owlish
This was previously disabled on release and beta builds in order to
reduce the binary size, but the inability to retrieve logs from users
is a hindrance to fixing bugs. Since we are now shipping AABs can
afford the increase in size.

Differential Revision: https://phabricator.services.mozilla.com/D212268

UltraBlame original commit: 5a3c143a268285c7ca8c04849e90ce991d3b72bb