Bug 1898745 - P4. Remove active section and active field related code in formautofill r=credential-management-reviewers,NeilDeakin
The new architecture no longer has the notion of an "active" section.
When autocomplete is triggered, we pass the "element id" of the triggered element and
use this id to indicate which form we are currently processing.

Because of this change, we no longer need the code that deals with an active section, field, etc.

Differential Revision: https://phabricator.services.mozilla.com/D211553

UltraBlame original commit: d0d65cb44bc76559fa6cf292df9c9e2a21db28b7