Bug 1898745 - P7. Do not delay identifying autofill fields after an input is focused r=credential-management-reviewers,NeilDeakin,issammani
This patch removes `setTimeout` in `doIdentifyAutofillFields` because it causes the
formautofill module to mark a field as autocompletable AFTER the autocomplete controller
has already started searching for autocomplete entries. The issue with calling
`markAsAutoCompletable` after searching has begun is that it resets the autocomplete
search state, which results in the autocomplete popup not being displayed.

Upon reviewing why we added the setTimeout in the first place (see Bug 1364818),
I believe it can be removed. Listening for the `DOMContentLoaded` event should
be sufficient for handling the autofocus case.

Differential Revision: https://phabricator.services.mozilla.com/D214275

UltraBlame original commit: 8daa01aff9d2ee8f929c36beb8642ea14b5eeab0