Bug 1874022 - pt 14. Seperate arena and non-arena code paths in PHC r=glandium
The Arena ID is wrapped in a Maybe, the construction of Nothing() shows up
in profiles.  By testing the maybe we show the compiler that it can optimise
away the construction of Nothing when inlining this code.

Differential Revision: https://phabricator.services.mozilla.com/D207680

UltraBlame original commit: f72b2832f50fb56b4dc4de970a27abe66ae697c3