Bug 1898153 - wasm: Generate all tier-independent stubs together in ModuleGenerator. r=jseward
Generate all of the stubs that don't depend on any tier-specific data in
one chunk at the beginning of the CodeBlock. The tier-specific stubs (entry
stubs) are generated separately at the end of the CodeBlock.

The purpose of this is to prepare ModuleGenerator for generating tier
-independent data into a separate CodeBlock from the rest of the module
code.

Differential Revision: https://phabricator.services.mozilla.com/D213151

UltraBlame original commit: 0d18640a68ae2fcd8f7aebe7471c907bd80228a4