Bug 1905356 - Increase number of frames that we track damage regions for. r=gfx-reviewers,gw
We use the EGL_KHR_partial_update and EGL_EXT_buffer_age extensions in
order to reduce the number of pixels rendered when only part of the
screen has been invalidated. Currently we keep track of the previous 2
frames' damage regions, meaning we are able to partially render when
the buffer age is 3 or less.

However, on Android devices with high refresh rate screens it is
common to encounter a buffer age of 4 or 5. This patch therefore makes
us track the previous 4 frames of damage, meaning we are able to
partially render again on such devices.

Differential Revision: https://phabricator.services.mozilla.com/D215206

UltraBlame original commit: 4ed9bc048396ed2e5371317a1a846313820ce0b0