Bug 1900070 - [devtools] Mark invalid at computed-value time declarations in computed panel. r=devtools-reviewers,ochameau.
On the server, for custom properties, we check if they're registered, and in such
case, return a property to indicate if they're invalid at computed value time.
On the client, for the properties that are IACVT, we display the same error icon
that we show in the rule view when a declaration is IACVT, and indicate in the
title what's the expected syntax.
The test we already have for custom properties in the computed panel is expanded
to check that we display the error icon as expected.

Differential Revision: https://phabricator.services.mozilla.com/D215319

UltraBlame original commit: 4bf4bc201115729660ea73c22b728ba9d6833020