Bug 1892514 - part 5-2: Add optional assertions into `IMEContentObserver::FlatTextCache` to detect creating invalid cache r=smaug
It's hard to write text change notification tests to verify the cache in
`IMEContentObserver`.  Therefore, assertions helps to detect the bug of the
cache with existing automated tests.  On the other hand, the verification
requires a lot of CPU time if editor inserts a lot of text, etc.  Therefore,
I'd like to add assertions behind a pref and enable it in some specific tests
and entire the tests of editor/editing API.

Differential Revision: https://phabricator.services.mozilla.com/D211834

UltraBlame original commit: 6ed15cfce6df1e2822f08519d76e399e6a10c208