Bug 1897444  - Pass base color in add_composite_prim, then ignore it in the shader. r=gw
This appears to work around the glitch on Windows + Adreno. It is unclear why. The fact that this so far only reproduced on windows on arm looks like a driver issue, however from the workaround it looks like there might be a confusion between the textured and non-textured code paths with composite quads.

Differential Revision: https://phabricator.services.mozilla.com/D215622

UltraBlame original commit: 9ac7f1f672d585d3a66062be012715bfafbb1e4d