Bug 1898665 - Fix crash in ExternalAppBrowserFragment r=android-reviewers,petru
ExternalAppBrowserFragment was resetting the content for the navbar (because the design for navbar inside a custom tab is different), but it didn't account for cases where the navbar feature was on, but the navbar would still be hidden (landscape/tablet). The patch sets the proper condition.

Differential Revision: https://phabricator.services.mozilla.com/D215711

UltraBlame original commit: 7b10b679535425b8fcfcc75f74cea73150cc7fef