Bug 1906078 - Avoid pathological VecDeque behavior on parallel traversal. r=dshin
Using split_off to split the front chunk from the queue seems like it'd
be efficient, but it's not:

  https://github.com/rust-lang/rust/issues/127281

Instead, pass the range of nodes to distribute, and use
VecDeque::from_iter in the callee. Then, just truncate() the queue to
the remaining local work.

Differential Revision: https://phabricator.services.mozilla.com/D215705

UltraBlame original commit: 9596795522df50332e495b900ac50b85e2f431f3