Bug 1892684 - [devtools] Fix the text indent when the conditional panel is open r=devtools-reviewers,nchevobbe
The story here :-
CM6 adds a padding of 6px as a space between the gutter and the editor content.
The consequence of this the conditional panel will have a 6px gap on the left side.

Prior to this patch i removed the padding on the .cm-line element which contains the
conditional panel, this causes the content on the line to shift and ..viola ..the BUG.

To fix this, i'm adding 6px `padding-inline-start` to compensate.

Differential Revision: https://phabricator.services.mozilla.com/D215269

UltraBlame original commit: 038c252a09532d07f5bec8dc2314e039b4578445