Bug 1906451 - wasm: Always return an AnyRef from CheckRefType. r=bvisness
CheckRefType used to return through one of two out-params depending
on what kind of reftype you asked for. This was not correctly handled
for typed references to functions. This is overall error prone for
little benefit, so this commit refactors this so that we always
return through an anyref out-param. Users who want a JSFunction*
can just grab that through that anyref easily.

Differential Revision: https://phabricator.services.mozilla.com/D216064

UltraBlame original commit: 210ec778dbfcfcc58107f62729b15e0769bfe649