Bug 1904048 - Add prototype for ImageBitmap's BindingJSObjectMallocBytes. r=aosmond
It looks like there was never a prototype for ImageBitmap's BindingJSObjectMallocBytes function.
Even though it was defined and expected to associate ImageBitmap's internally allocated surface
memory for the GC to track, the bindings generator never actually tried to call it because the
lack of the prototype caused the default non-specialized version of this function to be called
instead. With the protoype in place, this actually gets called and the GC is able to properly
clean out ImageBitmap's in a more timely fashion.

Differential Revision: https://phabricator.services.mozilla.com/D216250

UltraBlame original commit: 6b4223a9d1cf21905ef5a591398a69a41ca9387e