Bug 1907320 - Allow sharing styles that specify transitions / animations. r=firefox-style-system-reviewers,layout-reviewers,boris
Most styles that specify transitions don't actually have a transition
running at the very same time, and we can check for that in the cache
already.

We need to avoid reusing-by-rule-node for starting-style rules, but
that's fine because that's effectively what already happened before this
change, due to the check we're removing (we only resolve starting-style
when we're about to maybe-start a transition).

Differential Revision: https://phabricator.services.mozilla.com/D216286

UltraBlame original commit: cf8b2df49bf0901b08109fe6f4d5ba4c178c1a0f