Bug 1906350 - Avoid refcounting overhead for string buffers passed from DOM code. r=sfink
The `RefPtr<StringBuffer>` argument of the new JS APIs can have pretty significant performance
overhead due to the atomic refcount used for `StringBuffer`.

This patch adds separate APIs that only use `RefPtr` when we're going to use the buffer for
a new JS string. This is a more similar to the behavior with the old external-string code
and fixes most of the reported performance regression.

Differential Revision: https://phabricator.services.mozilla.com/D216285

UltraBlame original commit: 8bda0e83ceba4450938b5cdcce406bb49bb8c4ac