Bug 1907901 - Text Fragments: Separate json test case to let others pass. r=dom-core,sefeng
The test to check if scrolling is allowed in non-html documents times out for JSON,
because it depends on the `onload` event, which Gecko is not firing for JSON (Bug 1905106).

Additionally, other test cases in that file can time out on CI as well.

With this patch, the JSON test case is separated into its own file and is expected to `TIMEOUT`. The other test cases get a longer timeout so that they pass in all configurations.

Differential Revision: https://phabricator.services.mozilla.com/D216565

UltraBlame original commit: 72e40f9471c4f8348d84d2e211b97f26c4a84d7b