Bug 1901497 - Enable -moz- transform prefixes on nightly again. r=webcompat-reviewers,twisniewski
This changes the approach of zoom+moz-transform compat on Nightly, on
the basis that its easier to site-patch stuff that uses both
-moz-transform and zoom (to disable either of them via CSS) than working
around bogus UA sniffing or feature detection that is sorta-reasonable
(like assuming that if "MozTransform" is in the CSSStyleDeclaration
object, then `-moz-transform` works).

If breakage from this is too severe, or the hypothesis above doesn't
hold, then we need to re-evaluate this in some other way.

Differential Revision: https://phabricator.services.mozilla.com/D216840

UltraBlame original commit: c35050f1941c0cc4044c1a94aac83b8ec6970cc2