Bug 1908587 - Revert all cairo-quartz code (quartz-surface, quartz-font, quartz-image) to the version prior to the cairo-1.18.0 update. r=gfx-reviewers,jrmuizel
The cairo-1.18.0 version of the cairo-quartz backend is resulting in unwanted
rasterization of the output when printing GDocs to PDF (or to a real printer,
presumably).

It's likely this would affect other canvas-based content as well.

The problem in the 1.18.0 backend probably relates to failing to use CGLayer-
backed surfaces in certain cases, but will require more extensive debugging
to fully understand and resolve.

Given that this worked better prior to the cairo-1.18.0 update (using the
old cairo-quartz backend that included a number of mozilla-specific adaptations,
including CGLayer support), the simplest/safest mitigation in the short term
seems to be reverting the entire quartz backend. Fortunately, it is sufficiently
isolated from the rest of cairo that this can be done without touching any
other cairo code outside the macOS-specific cairo-quartz-* files.

Differential Revision: https://phabricator.services.mozilla.com/D217012

UltraBlame original commit: 05bf7d6afa7d52303f68a26f8bba206f049c13e7