Bug 1908545 - Fix uncaught error when detect language is selected in about:translations. r=translations-reviewers,nordzilla
An earlier change updated `this.translator` to be a `Translator` object
instead of a promise that returned one. This patch fixes a line missed
in the original update, which caused erros to be thrown as `.then` is
not a valid method of `Translator`.

Differential Revision: https://phabricator.services.mozilla.com/D216875

UltraBlame original commit: 9eaa5e0d909a7a9541e61b53844de00374aa8ac8