Bug 1892762 - part 3 - Ensure snackbars for deleted bookmarks are shown above (vertically) the browser navbar r=android-reviewers,gl
The snackbar for deleting a just added bookmark is shown from the bookmarks screen which does
not have access to browser's toolbar to anchor the snackbar to it.
We can reuse the existing SnackbarBinding to update it about needing to show the snackbar
which will use a snackbar delegate with access to the needed anchor.

Differential Revision: https://phabricator.services.mozilla.com/D215908

UltraBlame original commit: 705ed5a56fcd77249a00fb0c86d6fd3e336d753c