Bug 1900846: Pass each individual field as parameters instead of the CycleCollectorResults struct when emitting the CCIntervalMarker profiler marker. r=canaltinova
The CCIntervalMarker marker accepts a CycleCollectorResults struct as one of it's parameters and emits each field individually during StreamJSONMarkerData.  This is difficult to translate to a perfetto track event, so instead pass each field directly to profiler_add_marker so that the parameters line up with the PayloadFields descriptions.

Differential Revision: https://phabricator.services.mozilla.com/D217411

UltraBlame original commit: 3a9a742747b0e749702883b06e408da22663eb48