Bug 1907945 - Use i64 instead of u64 in custom_distribution ctor r=TravisLong
This makes me sad, since these are for quantities that only permit unsigned
values. But this is what the Glean SDK was forced to use by early Kotlin,
and it simplifies codegen a little now that the signed types are re-exposed
by LabeledMetricData... so I guess we suffer.

Differential Revision: https://phabricator.services.mozilla.com/D217606

UltraBlame original commit: 99c22c70694e39207641f6c24632c5477c890405