Bug 1908470 - Make it harder to detect when PPA is disabled r=emilio
The privateAttribution API returns early when the user has disabled
Privacy Preserving Attribution. This makes it possible to detect this
condition by passing in an invalid target domain and observing whether
the API raises an exception.

This moves the user preference checks to happen after input validation,
so that input validation happens (and an exception is thrown) even when
PPA is disabled.

Note that it might still be possible to detect whether PPA is enabled by
observing the timing of calls.

Differential Revision: https://phabricator.services.mozilla.com/D217642

UltraBlame original commit: e8ca6c27ae5bcfb694ba7228e23a5a84b510371e