Bug 1911152 [wpt PR 47409] - [CompositeClipPathAnimation] Remove HasCurrentClipPathAnimation, a=testonly
Automatic update from web-platform-tests
[CompositeClipPathAnimation] Remove HasCurrentClipPathAnimation

This CL fixes a CHECK failure that occurred on multiple websites
resulting from the HasCurrentClipPathAnimation flag being unexpectedly
unset when kNeedsRepaint was set following an animation being set
compositor pending. This situation arises when a transition completes, and also in certain other scenarios that are difficult to debug. In these cases, a compositing decision should still be made, however the flag is not set due to logic in
CSSAnimations::UpdateAnimationFlags. As such, the issue lies with the
CHECK - its failure in this scenario is not indicative of a code fault.
This check was added to ensure that kNeedsRepaint wouldn't persist
beyond the lifetime of an animation, however the effect of delays and
suppressed animations were not considered.

The HasCurrentClipPathAnimation flag was initially added to avoid heavy logic in HasCompositeClipPathAnimation being run every repaint.
However, this was recently ameliorated in favor of using the new
kNoAnimation state for the composited paint status, so this flag is now unused except as a CHECK.

Rather than attempting to align these two values, this CL simply removes HasCompositeClipPathAnimation.

Bug: 355108923
Change-Id: Iaa57ad84212ac1fdcd8fb231685697cafb41ea91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5737832
Reviewed-by: Kevin Ellis <keverschromium.org>
Commit-Queue: Claire Chambers <clchambersmicrosoft.com>
Cr-Commit-Position: refs/heads/main{#1336131}

--

wpt-commits: 134837b5dee26f2ff29e1455d11ed6011ce39e7f
wpt-pr: 47409

UltraBlame original commit: 8af95efb377ab715ac80747d5f995792311d20d3