Bug 1858740 - Use ARM builder for Mac CaR builds. r=perftest-reviewers,taskgraph-reviewers,ahal,sparky
This patch switches the builder from the intel osx 10.15 builders to the
Apple Silicon builders. 10.15 Is no longer a supported build target for
CaR (Chromium-as-Release) so let's use this opportunity to just go to the latest version
builders we have.

Additionally, this patch removes the custom CaR osx 10.15 tasks as that
is no longer a supported platform for the application.

Differential Revision: https://phabricator.services.mozilla.com/D218368

UltraBlame original commit: 6582643381c93098c53e8dc62d8de1ca27bf2814