Bug 1911773 - Simplify nsWindow::SetWindowTranslucencyInner. r=rkraesig,win-reviewers
We no longer have the parent/child window set-up for popups, so we can
just use mWnd.

Remove the HideWindowChrome() stuff. That landed once already in [1],
but was backed out along with other stuff, due to it causing buttons to
show up on non-native notifications (which is the only top level that
hits this code-path).

Eventually, I want to handle this more like other platforms (via D209502
and D207303) but until those land, just explicitly set hidechrome="true"
on that window in order to be able to land this simplification
separately.

Bug 1454589 doesn't happen with this change, and there's no reason to
apply size mode in SetWindowTranslucencyInner, so remove that chunk of
code too.

[1]: https://hg.mozilla.org/mozilla-central/rev/24dc60b32e21

Differential Revision: https://phabricator.services.mozilla.com/D218635

UltraBlame original commit: 253ac2ff1e9e35dde102f3cabfeb58710868a52c