Bug 1910242 - Fix crash reporting view top and bottom margins with navbar on r=android-reviewers,petru
Previously, the CrashContentIntegration would consider the position of the toolbar and set the marging to the corresponding side. With the navbar we might need to set marging on both sides, and might need to adjust it on the orientation change.

Differential Revision: https://phabricator.services.mozilla.com/D218074

UltraBlame original commit: d4dc6bfb438ccd97d09cf0c633bdccb076b5ca8f