Bug 1912328 - Permit stat for allowed files in GMP process sandbox. r=gcp,gerard-majax
After updating the Widevine plugin to 4.10.2830.0, we would crash on
startup of the plugin because it attempted to use the stat syscall.
Allow uses of stat for files that we have already opened / allowed
access to in the GMP sandbox.

Differential Revision: https://phabricator.services.mozilla.com/D218855

UltraBlame original commit: 5a91437d9dc5787cab9cef1038c3073d85226a76