Bug 1913311 - Add a testcase r=media-playback-reviewers,aosmond
This patch introduces a test that demonstrates a crash occurring when
`DecoderAgent::Shutdown` is called without an active `MediaDataDecoder`.
The issue arises when `PDMFactory::CreateDecoder` in
`DecoderAgent::Configure`, rejects decoder creation, leaving
`DecoderAgent::mDecoder` as `null`. Subsequently, when
`DecoderAgent::Shutdown` is invoked, the assertion that checks for the
existence of `mDecoder` fails, leading to a crash.

To replicate this scenario, the test includes a pref that forces the use
of a `NullDecoderModule`, and ensures that this module rejects decoder
creation.

Differential Revision: https://phabricator.services.mozilla.com/D219709

UltraBlame original commit: 6a1aef3463658615f595d5b2b5e051ec81af5b2d