Bug 1914831 - Try to use consistent rounding in DOMRect::SetLayoutRect. r=longsonr
Otherwise we can get into states where identical numbers end up being
different from the ones exposed in other APIs.

This is a somewhat risky change, so gate it behind a pref.

This fixes the issue, but given it's a tricky change I'll also add a
workaround to avoid depending on it on the front-end.

Differential Revision: https://phabricator.services.mozilla.com/D220126

UltraBlame original commit: 75588bc186227bd022a84770ec16be91cff3f0cc