Bug 1915014 - Remove fallback to hidden window in ParentProcessTargetActor._getInitialDocShell r=ochameau,firefox-desktop-core-reviewers
This was added in bug 1275942, which added the --wait-for-jsdebugger option.
I think it's not necessary because BrowserGlue.sys.mjs is loaded before
"command-line-handler" category runs[1], so the early blank window will
have already opened before DevToolsStartup.handleDebuggerFlag spins the
event loop[2].

Currently, the early blank window is not enabled on all platforms and
can be disabled by pref. To make it reliable for DevTools use, we will
forcefully create the early blank window regardless of OS or pref when
--wait-for-jsdebugger is passed to the application. I tested this patch
on macOS (where the early blank window is not enabled) and it works well.


[1] https://searchfox.org/mozilla-central/rev/3f193a8bb0759058c2f9a9d7761ad73bc6ffd4b1/toolkit/xre/nsAppRunner.cpp#5571-5573,5734-5737
[2] https://searchfox.org/mozilla-central/rev/1d4c27f9f166ce6e967fb0e8c8d6e0795dbbd12e/devtools/startup/DevToolsStartup.sys.mjs#1011,1037-1043

Differential Revision: https://phabricator.services.mozilla.com/D220171

UltraBlame original commit: e04c427060e192c073a8005457532e6e62be9042