Bug 1915111: Double clicking in the vertical tab strip should open a new tab. r=desktop-theme-reviewers,tabbrowser-reviewers,sclements,dao
This solves two problems. First we must make the tabbrowser-tabs instance
fill the remaining space in the sidebar while at the moment it is constrained
to the height of the tabs. Then we allow the double click event to be
handled when vertical tabs and tabs in titlebar are enabled.

Differential Revision: https://phabricator.services.mozilla.com/D220269

UltraBlame original commit: 7f290c5b5b1aa90622da3630574b3687fd5a3e3b