Bug 1913403 - revert bug 1911651. r=media-playback-reviewers,padenot
While cbsnews.com uses hls.js on desktop, it fails to do so for Fenix.
Since it's a major news source in the US, the impact will be huge.
Reverting the change to not break the site for release.

Differential Revision: https://phabricator.services.mozilla.com/D220814

UltraBlame original commit: c15cb2fe71dbb5334995f55355c20fc181ae4e08