Bug 1907926 - Add an local alert test run layer to mozperftest. r=perftest-reviewers,aglavic DONTBUILD
This patch adds a new layer to mozperftest that allows people to run all the tests that produced an alert locally.

It does this by querying Treeherder for the alert information, and determining the tests to run from there. A `--alert-exact` option is added to enable users to run the exact same command as was run in CI as well. This is done by querying the task information of the alerting task, and parsing the options from it in a framework-specific manner.

All harnesses have option parsing implemented for them, however some of them, such as AWSY are more manual due to how the argument parser is implemented in that harness.

Differential Revision: https://phabricator.services.mozilla.com/D219965

UltraBlame original commit: 6c3f81917e3407307a9d129ea4d11ee873bbc834