Bug 1880582: Use more correct teardown semantics in nsCocoaWindow::Destroy. r=mstange
This patch moves the call to BaseWindow::Destroy() and ::OnDestroy() to
the end of nsCocoaWindow::Destroy(). It adds a death grip on this before
remove this from mParent. This should keep the window in a more stable
state throughout the function scope.

Differential Revision: https://phabricator.services.mozilla.com/D216066

UltraBlame original commit: 90950f99d69f15ec31fcedced1ac914e761baf19