Bug 1914947 - Add the necessary tools to build clang on mac hosts. r=firefox-build-system-reviewers,sergesanspaille
While we are building cmake from source on Windows, that's only because
we're patching it to work around a cmake issue on Windows. We used to
use a binary on Windows before, so we don't shy away from using a binary
on mac.

Differential Revision: https://phabricator.services.mozilla.com/D220528

UltraBlame original commit: c35b28d242ff3134f28de8fff2293f10394b050d