Bug 1916008 - Separate signal handling for AudioTrimmer and MediaChangeMonitor creation r=media-playback-reviewers,padenot
Previously, `AudioTrimmer` and `MediaChangeMonitor` relied on the same
signal for their creation since `AudioTrimmer` was used exclusively for
audio and `MediaChangeMonitor` for video. However, changes introduced in
the patches for Bug 1891082 will enable MediaChangeMonitor to work with
audio as well. To allow for the possibility of wrapping an
`AudioTrimmer` within a `MediaChangeMonitor`, these components now
require separate signals for their creation.

This patch replaces the `NoWrapper` class, which was previosuly used to
signal the creation of both `AudioTrimmer` and `MediaChangeMonitor`,
with a new `Wrapper` class that includes two bit flags:
`Wrapper::AudioTrimmer` and `Wrapper::MediaChangeMonitor`. This allows
the creation of `AudioTrimmer` and `MediaChangeMonitor` to be controlled
independently, or together, by using `Wrapper::AudioTrimmer`,
`Wrapper::MediaChangeMonitor`, or `Wrapper::AudioTrimmer |
Wrapper::MediaChangeMonitor` respectively.

Depends on D220965

Differential Revision: https://phabricator.services.mozilla.com/D220753

UltraBlame original commit: fef9517a08aee55c417691787ec454611b15d43e