Bug 1912988 - part 2 - Ensure we show a divider only when the microsurvey is showing r=android-reviewers,skhan
This ensures that the measured height of (toolbar + navbar) - used in the setVerticalClipping
Call will be the same as the computed height of (toolbar + navbar) used in the
setDynamicToolbarMaxHeight call.

Will also bring a more cohesive UI between the home and browser screens which is always good.

Differential Revision: https://phabricator.services.mozilla.com/D220867

UltraBlame original commit: 63b1685c86c6861fb2d04f2112643dd30e29510b