Bug 1885446 Part 1: Stub the WebGL unpackColorSpace attribute. r=jgilbert,webidl,smaug
This is a stub so webidl reviewers can review this without having to review
the implementation. This is not attempting to do anything with
unpackColorSpace other than ensure that it is get and set in the same
ways that are used for drawingBufferColorSpace.

Differential Revision: https://phabricator.services.mozilla.com/D214053

UltraBlame original commit: b462660b4d141e5ca071f2de8c798541c8511111