Bug 1917496 - Update AwesomeBarFeature.onCancelEditing based on editing state changes r=mcarare,android-reviewers
The `AwesomeBarFeature` has an `OnEditListener` that always returned
the callback for `onCancelEditing` as true. This resulted in the
`Toolbar` updating it's internal state from `State.EDIT` to
`State.DISPLAY` when no change had happened.

This patch uses the existing state variable to respond to the
callback which mimics the correct behaviour of the AwesomeBar.

Differential Revision: https://phabricator.services.mozilla.com/D221416

UltraBlame original commit: 3f2359e25162af006529cc41c04f3c96ee0945fb