Bug 1583641 - [devtools] Handle pseudo-element rules of selected node in Inactive CSS. r=devtools-reviewers,ochameau.
The inactive-property-helper can sometimes be called for properties inside rules
that match pseudo-elements, even though the node being passed is not the pseudo-
element itself.

This happens because the Rules view has a feature whereby selecting, say, a <div>
will show its rules, but also any pseudo-element rules that may be applied that
result in the creation of a pseudo-element inside of it.
So if you have div::before { content: "" } in CSS, then selecting that div will
also show this rule in the Rules view.

This patch adds a check for this case, and when it happens, walks the DOM to
find the right pseudo-element, and then gets its style.

Differential Revision: https://phabricator.services.mozilla.com/D65720

UltraBlame original commit: 66ecda90e9314227cda14714ccb677a30408eba0