Bug 1918371 - Use browser-custom-colors.css background if appropriate. r=desktop-theme-reviewers,dao
Remove the sidebar-border variable because it's unused (and it's not one
of the properties that get plumbed down anyways).

Rely on browser.xhtml's background to avoid having to plumb down the
system theme colors. Alternatively we could try to plumb these down, or
add a media query like:

```
media not ((prefers-contrast) or (-moz-platform: linux) or ((-moz-platform: macos) and (-moz-bool-pref: "browser.theme.macos.native-theme"))) {
  /* Duplicate browser-custom-colors.css */
}
```

But that's kind of unwieldy, to the point that we should move
browser-custom-colors.css import to browser-shared and use the same if
we do it, probably, so it's easy to keep them in sync...

Differential Revision: https://phabricator.services.mozilla.com/D221962

UltraBlame original commit: fd7ccc8b9088878e6c33cfc587929ffb9e6c1a11