Bug 1917458 - Only send resizes for remote browsers once per frame. r=mstange,extension-reviewers,tabbrowser-reviewers,dao,robwu
Do it at the same time as UpdateRemoteFrameEffects() (so, intersection
observer timing). Otherwise we can see flickering sometimes from the
resize with the previous patch (and this should be less work, when we
resize).

Non-remote frames need to synchronously communicate their resizes
(because JS could access the frame), but that's not an issue for
remote frames.

This should also more reliably prevent issues like bug 1910887 or like
bug 1764655, and paves the way for fixing bug 1750189 (which stalled) in
a similar fashion.

I tested this in a build with a couple hundred tabs open and it doesn't
measurably show up. I think we should consider not communicating resizes
to background tabs / hidden remote iframes, at least for top levels.

Differential Revision: https://phabricator.services.mozilla.com/D221405

UltraBlame original commit: 65384dcd3d9a1383df9f46fb70e8c92ce7e69df1