Bug 1917467 - Remove mozplugin-block-digest256 from the urlclassifier.blockedTable for Beta and Release. r=dimi
The test test_initial_download.py is going to fail in beta and release
channel because we only populate the mozplugin-block-digest256 in the
nightly channel. To fix this, we remove the mozplugin-block-digest256
from the urlclassifier.blockedTable pref for Beta and Release channel to
exclude testing mozplugin-block-digest256.

Differential Revision: https://phabricator.services.mozilla.com/D222004

UltraBlame original commit: a5b38f565e804b0f6092cc78e4afd742099a5122