Bug 1917862: Part 1- Updated collection composable so that when the expanded value is remembered, it uses the expanded value as a key in addition to the collection itself. r=gl,android-reviewers,007
With the current RecyclerView implementation, I think this is working because it's passing in a new collection object every time the expanded value changes. I updated it so that it'll register the change if the expanded value is changed.

Differential Revision: https://phabricator.services.mozilla.com/D221723

UltraBlame original commit: 83e3e2b93565bd68fbef5617a3d51d0b9da2b3b1