Bug 1916333 - [devtools] Adapt console filter input colors in High Contrast Mode. r=devtools-reviewers,accessibility-frontend-reviewers,nlapre,ochameau.
This adds proper border and outline colors to the console filter input.
The filter height is reduced so those are visible, and we also
hide the separator around it which are not needed in HCM since we
display the border on the input.

Differential Revision: https://phabricator.services.mozilla.com/D222069

UltraBlame original commit: 76364a160d4d3288e34a9625a7a2b4f73996e580