Bug 1911071 - part 2: collect and print statistics on lazy-tiering and inlining.  r=yury.
This patch adds machinery for collecting statistics on lazy tiering and
inlining.  These are stored in the CodeMeta for a module and are printed, using
the MOZ_LOG facility, when the CodeMeta is destructed (at the end of the
module's lifetime).  The collected data is minimal but is important for tuning
the LT and inlining heuristics.

* WasmMetadata.{h,cpp}, class CodeMetadata
  - new struct ProtectedStats and lock-protected field `stats`
  - new method CodeMetadata::dumpStats

* vm/Logging.h: new logging tag 'wasmCodeMetaStats'

* vm/MutexIDs.h: new mutex 'WasmCodeMetaStats' for serializing access to the
  stats

* WasmCode.cpp: copy code-allocation info into CodeMeta::stats

* WasmGenerator.cpp, ModuleGenerator::finishModule: copy complete-tier info
  into CodeMeta::stats

* WasmIonCompile.cpp, class FunctionCompiler:
  - add comment explaining how the FunctionCompiler stack is organized
  - new field `toplevelCompiler_`, to access the stack base quickly
  - new fields `stats_*`, temporarily holding stats during compilation of one
    function
  - FunctionCompiler::~FunctionCompiler: copy stats into CodeMeta::stats

Differential Revision: https://phabricator.services.mozilla.com/D221936

UltraBlame original commit: 5b7ee12ee96d020a2e23c7fe785b3b54507095a0