Bug 1918907: Don't clear drag session target BrowserChild on eDragLeave/eDragExit r=m_kato
The BrowserChild should only be cleared on EndDragSession, which is sent when
the user finishes or cancels dragging. This is reciprocal to it being set in
StartDragSession.

The change to nsDragSessionProxy::EndDragSessionImpl is non-functional -- it
highlights the BrowserChild symmetry above.

Differential Revision: https://phabricator.services.mozilla.com/D222389

UltraBlame original commit: 6e042c75c31a96990f3a5dcede13fe6e478a8eff