Bug 1920055 - Use MODE_TRRONLY in test, r=necko-reviewers,valentin
The reason of this test failure is that bug 1918928 added a check to only wait for HTTPS RR when TRR is confirmed. However, in this test, the TRR is in mode 2 and it's not confirmed yet, so this test failed.
About the reason that we didn't see this in nightly is that HTTPS first mode is enabled by default, so the request is upgraded.

Differential Revision: https://phabricator.services.mozilla.com/D222942

UltraBlame original commit: 3c2d3da2aab546671ac2ab3629cdb71f825906a7