Bug 1919604 - Run tp6 pageload benchmark on only win/linux. r=perftest-reviewers,sparky
the `tp6-bench` works for the most part on windows and linux, so this
patch enables it to run on autoland so that data can start getting
collected. Additionally, keep this on tier 3 for now until we know more
about the data, and ignore osx 1015 & 1400 for now.

Differential Revision: https://phabricator.services.mozilla.com/D222685

UltraBlame original commit: dcc7831d7774cd90f0ead40dac6c5645d1c828b4