Bug 1919537 - don't crash with DLP on when drag/drop text to a site on Deny list r=dlp-reviewers,reusable-components-reviewers,handyman,mstriemer
This prevents a problem where Content Analysis responds synchronously (because
the site is on the deny list), which causes a modal window to pop up, which
cancels the drag session. This causes MOZ_ASSERTs() where we expect the drag
session to still be active.

Differential Revision: https://phabricator.services.mozilla.com/D222668

UltraBlame original commit: 01542f2b7bb2283115eb266bc1f380ee1340d77b