Bug 1920486 - Fix systemBlockedPermissionsRedirectToSystemAppSettingsTest UI test on API 34 r=aaronmt
The UI test was failing when trying to verify the permission summary in the system settings. (on APIs > 30 the summary isn't displayed)
To overcome this problem, I've added a condition to verifySystemGrantedPermission.
Left the min SDK set to 29, bellow that the system settings are significantly different.

The UI test:

  - Successfully passed 25x on API 30 ✅
  - Successfully passed 50x on API 34 ✅

Differential Revision: https://phabricator.services.mozilla.com/D223124

UltraBlame original commit: c9c370c83cb2dcc0efbe7358599b30038135db0f