Bug 1917992 - Add scalar telemetry for picked restrict keyword result and glean telemetry for abandonment and engagement. r=mak,urlbar-reviewers
This patch adds telemetry for when the user sees or interacts with restrict keyword results.
Following is an overview of the interactions that will trigger the telemetry to be recorded:

When the user types `` and selects one of the restrict results, the scalar telemetry will be recorded:

  - `urlbar.picked.restrict_keyword_tabs`
  - `urlbar.picked.restrict_keyword_bookmarks`
  - `urlbar.picked.restrict_keyword_history`
  - `urlbar.picked.restrict_keyword_actions`

For example, we can select the tabs result in multiple ways for the telemetry to be recorded:
  - type`` and click on tabs result
  - type`` and arrow down to tabs result and press enter
  - type partial keyword `t` and press enter to autofill
  - type partial keyword `t` and press tab and then enter


When the user types `` and **does not** select one of the restrict results, the glean abandonment will be recorded:
```
    groups:
      "general,general,general,general,general,general,restrict_keyword,restrict_keyword,restrict_keyword,restrict_keyword",
    results:
      "search_engine,search_engine,search_engine,search_engine,search_engine,search_engine,restrict_keyword_bookmarks,restrict_keyword_tabs,restrict_keyword_history,restrict_keyword_actions",
```

When the user types `` and **does** select one of the restrict results, the glean engagement will be recorded:
```
    groups:
      "general,general,general,general,general,general,restrict_keyword,restrict_keyword,restrict_keyword,restrict_keyword",
    results:
       "search_engine,search_engine,search_engine,search_engine,search_engine,search_engine,restrict_keyword_bookmarks,restrict_keyword_tabs,restrict_keyword_history,restrict_keyword_actions",
    selected_result: `restrict_keyword_${category}`,
```

For the `selected_result` in engagement telemetry, the `category` variable will be `tabs, bookmarks, history, or actions`.

Differential Revision: https://phabricator.services.mozilla.com/D222355

UltraBlame original commit: 1e3fb966ffb11605b527f896d1af070fa7dfffe3