Bug 1906795 – Implement new landscape design for custom tabs r=android-reviewers,skhan,twhite
With the addition of the navigation bar in the portrait mode, we are changing the UI on rotation: mainly, buttons, that were part of the navigation bar, are being added to the toolbar instead, to save space.

This patch is adding navigation, share, menu and open in firefox as actions to the toolbar.

Differential Revision: https://phabricator.services.mozilla.com/D220207

UltraBlame original commit: 40833555896a46ed8b10389fac977df5678ab964