Bug 1917662 - Rework additional metrics parsing for support classes. r=perftest-reviewers,fbilt
This patch reworks the additional metrics parsing for support classes into a more generic method. It separates the metric parsing into multiple methods as well to make it simpler for other support classes to use them directly.

At the same time, these changes make better use of the saved data for building the measurements, and adds functionality that allows callers to specify if they want cold, or warm measurements (in case these need to be separate).

Some comments are modified as well to better reflect inputs, and what the methods are meant to be used for.

Differential Revision: https://phabricator.services.mozilla.com/D222178

UltraBlame original commit: 99e9d5cf35c0b9af74de627506fb1f5de382f181