Bug 1838298 - Custom Stylelint plugin to prevent color tokens from being used directly r=Standard8,frontend-codestyle-reviewers,desktop-theme-reviewers,dao
This patch lays the foundation for a Mozilla-specific Stylelint plugin that we can continue to expand on with new rules as needed. For the first rule I've implemented a simple check to ensure our color tokens are being used as intended.

Differential Revision: https://phabricator.services.mozilla.com/D223204

UltraBlame original commit: f4e625378c9e599eeeed83222bb7a12be4864fc2