Bug 1521450 - Enable per-monitor DPI scaling V2 in the Windows GUI r=gsvelto,glandium
This is quite an improvement on the quirks of the previous GDI scaling.

It also mostly supports the windows 10+ "Make text bigger" setting: it
reads the value from the registry (albeit at an unofficial location),
but doesn't register a key change listener to update the value if it
changes while the application is open. I think this is very, very likely
to be good enough; I will be surprised if someone notices this
deficiency! The official API is part of UWP and is accessible through
C++ libraries, but not conveniently through win32 APIs, which is why I
use the registry.

Differential Revision: https://phabricator.services.mozilla.com/D221544

UltraBlame original commit: ab928446525345cb9f8b7e29c0497bcb1287d978