Bug 1921549 - List the keywords in the results panel when a user types ``. r=mak,fluent-reviewers,urlbar-reviewers
This patch adds the keywords to the result title for TokenAliasEngine and
RestrictKeyword results. The TokenAliasEngine results may have more than one
keyword if the user has set a custom keyword for that engine. This can help
remind the user what keywords are available as they are typing or if they
have forgotten a keyword, they can type `` to quickly see the list of keywords
to search with.

The result title looks like this:

For Engines:
 [keywords] - Search with [Engine]
For Restrict Keywords:
[keyword] - Search [Restrict]

Differential Revision: https://phabricator.services.mozilla.com/D224017

UltraBlame original commit: 90d40d9da4bcf102532f193010d731d78083d73c