Bug 1878837 - Fix verifyUBlockWorksInNormalModeTest UI test r=ohorvath
The UI test was flaky and even failed when trying to verify he number of blocked trackers.
Taking a closer look, I've noticed that a yellow badge with a "!" sign being displayed in the "Extensions" submenu. {F8336296}
I've managed to find the meaning of this on the [[ https://github.com/gorhill/uBlock/wiki/Toolbar-icon | uBlock wiki ]] and it seem to mean that "some network requests were fired by the browser while uBO was not ready to filter properly, potentially leading to ads/trackers/etc. not being filtered in some of the already opened webpages"

To overcome this problem I've added a retry loop when trying to verify the number of trackers blocked in case the above mentioned inconsistency is being encountered.

The UI test successfully passed 100x on Firebase on both API 30 and 34 ✅

Differential Revision: https://phabricator.services.mozilla.com/D224597

UltraBlame original commit: 89ee8693f2b51b2efbfab479d32eb25489dad6a8