Bug 1921200 - [devtools] Stop using hooks to call createValueGrip. r=devtools-reviewers,nchevobbe
For this, we have to be more explicit about the current depth and pass it
from each callsite.

I've made depth be a mandatory argument and pass a lot of 0's in order
to help spot cases where we are trying to spawn new root previews.

Differential Revision: https://phabricator.services.mozilla.com/D224113

UltraBlame original commit: 9b1af84f3f936672bf717557593cf5aceed5b0b3