Bug 1922645 - Do not use stale cached profile when authType is a new signin r=lina
This patch is a workaround for a bug where we can end up persisting
the profile for the previously signed in account during the
onAuthenticate because we haven't fetched a fresh profile when
authenticating.

Bug 1915630 is a better fix for that problem, but until then,
we workaround this by choosing to fetch a profile when the `authType`
is not an existing account.

Bug 1922650 is another follow-up to make bugs like this less likely
to occur.

Differential Revision: https://phabricator.services.mozilla.com/D224535

UltraBlame original commit: 443ae343f1ada4103f7e7e27767de3375af06114