Bug 1918053: Part 1: Check attributes when inserting a new StencilModuleRequest. r=jonco
When inserting a new entry in moduleRequests_, add a HashMap (moduleRequestIndexes_)
to check if there's an existing entry, if found, reuse the index of the existing
entry.

So now a ModuleRequestIndex will represent a unique index for a
(specifier, attributes, firstUnsupportedAttributeKey) tuple.
(Originally, a new ModuleRequestIndex will be created even the two
requested module have the same
(specifier, attributes, firstUnsupportedAttributeKey) tuples.)

And replace the requestedModuleSpecifiers_ with requestedModuleIndexes_,
which is a HashSet of the ModuleRequestIndex, to know whether to create a new
entry or reuse an existing entry in  moduleRequests_.

Differential Revision: https://phabricator.services.mozilla.com/D221804

UltraBlame original commit: c0f0c70cca1713e9bac219b606975e0ac062c8c9