Bug 1923204 [wpt PR 48506] - Add a WPT test validating a canvas filter with a globalAlpha, a=testonly
Automatic update from web-platform-tests
Add a WPT test validating a canvas filter with a globalAlpha

The specification currently says that the globalAlpha should be
applied on the result of the context's filter. This however leads to
an unexpected behavior because if the filter is a drop-shadow, we
can't see the shadow through the foreground even though the foreground
is made transparent by the globalAlpha.

This is not what Chrome and Firefox do. These browsers instead apply
the globalAlpha on the source of the filter, therefore making the
foreground transparent.

The specification issue is addressed in:
https://github.com/whatwg/html/pull/10674

Change-Id: I3832c3cf6390933770a08669f62e9b1b29caccf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5904448
Reviewed-by: Aaron Krajeski <aaronhkchromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravelchromium.org>
Cr-Commit-Position: refs/heads/main{#1365599}

--

wpt-commits: ac35e3fe543ee6266f40ed5a482bad80f7002e51
wpt-pr: 48506

UltraBlame original commit: c91aa8c57501776694dc4b4f7ba2a5c15e3d9188