Bug 1924276. Fix quadratic behaviour in nsIFrame::IsScrolledOutOfView. r=layout-reviewers,emilio
We walk up the frame tree looking for scroll frames, and then transform from the initial frame to the currently found scroll frame. The problem is that doing that transform walks between the initial frame and the current scroll frame again to get the transform, each time, and that operation is kind of slow even if we weren't doing all this extra work.

So instead we use the transformed rect and the current scroll frame so we don't have to compute the transform over those frames again as we walk up the frame tree.

Differential Revision: https://phabricator.services.mozilla.com/D225440

UltraBlame original commit: 092d5b1c73ba517f11a82e738e2e008e799e6d4a