Bug 1923081 - set define MOZ_SAFEBROWSING_DUMP_FAILED_UPDATES globally r=dimi
It is used to conditionally set a structure field, so it's required
that all compilation unit using it have the same value for that define.
The easiest way to achieve that goal is to set the define globally.

Differential Revision: https://phabricator.services.mozilla.com/D224752

UltraBlame original commit: 574660ebcc7d461ce480bd767e7620c12ae3f1e9