Bug 1813213 - Factor the dynamic toolbar bimodal height into position sticky elements stuck to the root scroll container. r=botond,geckoview-reviewers,m_kato
And kick a reflow for the sticky elements when the dynamic toolbar is hidden so that
the sticky elements' position laid out on the main-thread is properly updated there.

These stuffs are basically same as we've done for position:fixed elements.

Differential Revision: https://phabricator.services.mozilla.com/D223906

UltraBlame original commit: 67d2c05edf11b9e06ee35c49882a9aa677eaccea