Bug 1924888 - Swap order of arguments to DescribeScriptedCaller r=sfink
It's not a standard exception throwing JSAPI function. The one caveat here is
that I couldn't put JSContext in last position: There are default params to
preserve. So instead it's in second position, which still communicates that
it's not an exception throwing function

Differential Revision: https://phabricator.services.mozilla.com/D225726

UltraBlame original commit: ddea98533ca9c41538276501897136952f3c807d