Bug 1925808 - Part 2. Remove shmem usage from GMPVideoPlaneImpl. r=media-playback-reviewers,padenot
Now that we don't use GMPVideoPlaneImpl inside of GMPVideoi420FrameImpl,
there are no uses of GMPVideoPlaneImpl inside of our GMP implementation.
It is exposed as part of the GMP API, so we must continue to supply it,
but there is no need for it to be IPDL aware/performant.

Differential Revision: https://phabricator.services.mozilla.com/D226258

UltraBlame original commit: fb887a7aebf3794a9c56a5296aad2dbe34e87ffe