Bug 1925331 - Make the delete handlers of `HTMLEditor` guarantee that the point is in content node at calling `EnsureNoFollowingUnnecessaryLineBreak` r=m_kato
The point may not be set if the selection move is prevented or not suggested.
Therefore, it should check the caret position before calling
`EnsureNoFollowingUnnecessaryLineBreak`.  On the other hand,
`HandleDeleteCollapsedSelectionAtWhiteSpaces` should use `Selection` instead
if caret position is not suggested.  It should be handled in bug 1925424.

Differential Revision: https://phabricator.services.mozilla.com/D226092

UltraBlame original commit: 00271a7486f0a3874bdf784d2f876e552115d285