Bug 1925369 - Remove 'operating_systems' properties from Histogram definitions r=florian
All of the histograms were in code that only ran on the specified platform(s),
so the only thing that 'operating_systems' was doing was saving a few bytes
(the HistogramInfo, its part of gHistogramBucketLowerBoundIndex, its entry in
the PHF) and a few ms during build (less code in the headers,
fewer static asserts).

(( It didn't even save the bytes of storing the name in the string table. ))

Differential Revision: https://phabricator.services.mozilla.com/D226840

UltraBlame original commit: 581d74ca4495853f26ed6216a03fbff6564aa509