Bug 1927111 - Don't rely on static position (top: auto) for urlbar position. r=Gijs,sfoster,desktop-theme-reviewers,urlbar-reviewers,mak
The static position of these boxes is not great to depend upon, see:

  https://github.com/w3c/csswg-drafts/issues/9939

Right now, we are relying on the static pos to be properly set which is
going to go away, eventually, and it is bogus, because it doesn't update
if e.g. only transform changes.

Use getBoxQuads to get the content box bounds properly. Also, make
breakout position: absolute explicitly, to prevent confusion.

Differential Revision: https://phabricator.services.mozilla.com/D226932

UltraBlame original commit: 8afe68ed6509320c36f762f5fad05357f5504330