Bug 1925599 [wpt PR 48698] - Switch webkitgtk_minibrowser nightly runs to use a new universal bundle., a=testonly
Automatic update from web-platform-tests
Switch webkitgtk_minibrowser nightly runs to use a new universal bundle.

The runs on the CI (TaskCluster) of WebKitGTK have been not working since a
while [1] because the nightly bundles that we were shipping for download [2]
for testing on the WPT CI stopped working due to binary compatibility issues.

So we decided [3] that instead of shipping a bundle for a specific distro
we would ship an universal bundle that can work on any distro. This allows
us to ship a bundle with all the preview features enabled and the last
versions of all the libraries, matching the same configurations that are
tested on the release WebKitGTK bots at build.webkit.org

This commit changes the taskcluster init tasks, as also the installation
of the webkitgtk_minibrowser bundles to use this new universal bundle.

This also allows now developers to test with webkitgtk_minibrowser nightlies
independent of the distro they have. They simply need to pass the flags
"--install-browser --channel=nightly" to "wpt run"

[1] Fixes: https://github.com/web-platform-tests/wpt/issues/47823
[2] https://webkitgtk.org/built-products/x86_64/release/nightly
[3] https://bugs.webkit.org/show_bug.cgi?id=280523

--

wpt-commits: 9dc3d1b0f1c0c6571385ee41bd3cd6149d49ead8
wpt-pr: 48698

UltraBlame original commit: fe9601ab49ac33414a7eb1d02259f214cd259b97