Bug 1922506: Double deregister to ensure we don't leave any window actors. r=tjr
This can happen if the function that registered the window actor times out. Unregistering an actor that is already unregistered doesn't throw an error, it just logs it, so it is safe to call it twic an error, it just logs it, so it is safe to call it twice.

Differential Revision: https://phabricator.services.mozilla.com/D226046

UltraBlame original commit: 7e1b5399826cdefccbc858059b4ab22899fdfda2