Bug 1921759: Restart Windows Explorer at start of browser_bookmark_copy_folder_tree.js r=jmaher,win-reviewers,places-reviewers,rkraesig,mak
In a new Windows update, explorer.exe is grabs the clipboard mutex and holds
it for a long time, which is causing test failures.  Killing explorer.exe
restarts it and forces it to give up the mutex, allowing the test to pass.

Differential Revision: https://phabricator.services.mozilla.com/D227741

UltraBlame original commit: 6c79dac12a0e107e7da500317486018977ce9e8c