Bug 1916669 - [devtools] Don't use --theme-accordion-header-background on elements that are not accordion headers. r=devtools-reviewers,ochameau.
This adds a new variable with the same colors than the variable one that we can use as a drop-in replacement.
For some cases, we could reuse the existing --theme-body-emphasized-background.
The Debugger ResultList style could be simplified as it should always have either 'small' or 'big' class, which
were both setting a background color for selected item, overriding the declaration that was using --theme-accordion-header-background.

Differential Revision: https://phabricator.services.mozilla.com/D228148

UltraBlame original commit: 9a4b66196f172a4a713dd0bc571c6e5bbf94e62b