Bug 1930056 - Avoid empty tabstrip from shrinking. r=tabbrowser-reviewers,desktop-theme-reviewers,dao
In the situation in comment 0, the pinned tabs are absolutely positioned
until we recompute the position / fire the overflow event, which now
fires later. Thus, for a frame, we might have an "empty" tabstrip.

Fix this by ensuring that the tabstrip doesn't shrink under the minimum
height.

Rework a bit the variables to avoid repeating the math a third time.

Differential Revision: https://phabricator.services.mozilla.com/D228467

UltraBlame original commit: 138d80ec44f8f3caaa605f3190d6d8efc9173bb7