Bug 1927405 - Simplify and optimize OrderedHashTable::clear. r=jonco
Change `clear` to destroy all elements and then shrink the table only if needed,
similar to how `remove` is implemented.

This eliminates a `free` and `malloc` if there are only a few elements but it's also
simpler because it avoids calling `init` more than once which is a footgun.

Differential Revision: https://phabricator.services.mozilla.com/D227011

UltraBlame original commit: a07d438ea22ad946f59bf7244b77cd72074fce6c