Bug 1927464 part 7 - Store OrderedHashTable fields in reserved slots on MapObject/SetObject. r=jonco
This patch adds `OrderedHashMapObject` and `OrderedHashSetObject` base classes for JS
`Map` and `Set` objects, and separate `OrderedHash{Map,Set}Impl` templates operating on
these objects. Reserved slots are used for the hash table state.

This eliminates a `malloc` call for each `Map` or `Set` and a level of indirection.
Later we could delay the remaining `malloc` call until the first entry is added.

The `Range` objects no longer store a pointer to the hash table. These methods also take the
object as argument. This is a bit awkward, but in follow-up changes we can try to fold
`Range` into the iterator objects and add a separate `forEach` function to replace the
`Range` uses in C++ code.

Because the head of the `ranges` and `nurseryRanges` lists is now stored in a reserved slot,
`MapObject` and `SetObject` have an `objectMoved` hook to update the first range's prev pointer
when the JS object is moved.

`MemoryUse::MapObjectTable` is now used for the table's malloc buffer. A later patch will
rename it to `MapObjectData`.

Differential Revision: https://phabricator.services.mozilla.com/D227153

UltraBlame original commit: 145770c00ca8a5e77dc0cbcbf69de54ffb060f09