Bug 1927464 part 9 - Make remove and clear operations infallible. r=jonco
Shrinking the table is an optimization and JS code can already observe this state
if it catches the OOM exception, so it's simpler to ignore it.

The callers are still fallible due to `CallObjFunc`, but as a follow-up we can probably
change how that works and make `JS::SetClear` etc infallible too.

Differential Revision: https://phabricator.services.mozilla.com/D227155

UltraBlame original commit: 94af45f175fc617666ac3d07e9e3e7b71a4ee939