Bug 1929307: Import prefs from SharedPrefs at startup rather than clobbering them. r=jhirsch,niklas
This also stops sharing browser.profiles.enabled and toolkit.profiles.storeID. In the first case
I want to minimise any risk of this preference getting set to false by accident because then a
user loses all of their profiles. In the latter case there is no point storing it in the database,
we already have a backup mechanism for it and if we have lost the storeID we can't load the database
anyway.

Differential Revision: https://phabricator.services.mozilla.com/D228174

UltraBlame original commit: 260400cdb47fd265115e146caccb7bde955c80b3