Bug 1916700 - [devtools] Cleanup toolbarbutton.css. r=devtools-reviewers,bomsy.
No element with the `devtools-toolbarbutton` class can have a label or a checked attribute,
or can have `.toolbarbutton-text` items, so we can remove some selectors and rules from the stylesheet.

We take this as an opportunity to refactor the stylesheet with CSS nesting.

Depends on D221945

Differential Revision: https://phabricator.services.mozilla.com/D227838

UltraBlame original commit: b981cc8ac57224de2672920381ab4367b304617c