Bug 1930832 - Update history metadata observation behavior for App Services 134.2024111305025. r=android-reviewers,kaya
The latest version of App Services defaults to ignoring history
observations for pages that aren't in history. This is the
correct behavior for visit time observations, but not for
document type observations, which are recorded on page load,
before the page might have been added to history.

Differential Revision: https://phabricator.services.mozilla.com/D228765

UltraBlame original commit: 2de82f7026e8043c7b55db9a996a39163a5beb76