Bug 1930216 - Increase small heap incremental GC limit to improve splay benchmark behaviour r=sfink
This parameter changes the threshold and which we go non-incremental and the
threshold at which we consider GC urgent (which happens before that). It's the
latter limit we were previously hitting.

The exact limit is calculated based on heap size, which is categorized into
small, medium or large. This change affects the non-incremental limit for small
and medium heaps so should not increase the chance of OOM too much.

Differential Revision: https://phabricator.services.mozilla.com/D228475

UltraBlame original commit: 1d097cb528ec4e365e3d7500333489ba46742987