Bug 1930477 - Don't trigger slices on allocation while waiting for background unmarking r=sfink
This works the same as for other GC phases. Zone::setGCSliceThresholds sets the
slice threshold based on the result of GCRuntime::isWaitingOnBackgroundTask()
so as not to trigger slices every MB when we are waiting. We must trigger a
slice at the end of unmarking, which will enter the GC on the main thread and
reset these thresholds.

We also yield after the prepare phase if we were triggered by the background
task so as to let the embedding schedule the next slice in case this happens at
an inconvenient tim.

Differential Revision: https://phabricator.services.mozilla.com/D228594

UltraBlame original commit: f363f14f95d7fdd74835b93de5028dca20cf9ee2