Bug 1928666 part 1 - Replace OrderedHashTableImpl::all with forEach methods. r=jonco
This is a bit more ergonomic and faster. Ranges are now only used for the JS iterator
objects and this makes it easier to reason about them.

As a follow-up we could move the `Range` fields into reserved slots of the iterator
objects and avoid the separate allocation.

Differential Revision: https://phabricator.services.mozilla.com/D228978

UltraBlame original commit: d44f1938c3a2085929f9465f9540dcc13dcbbbf9