Bug 1931351 [wpt PR 49172] - Fix flakiness in touch-action WPT for deleted element., a=testonly
Automatic update from web-platform-tests
Fix flakiness in touch-action WPT for deleted element.

The test deleted the touch-element on pointerdown, and expected the
TestDriver action sequence to keep sending events at a position
relative to that element.  This assumes, without any supporting
documentation, that TestDriver determines the event positions before
the first dispatch.  My local testing shows otherwise: apparently
TestDriver "remembers" the target for latter dispatches in the
sequence, which is problematic after the target is deleted on
pointerdown.  This assumption made the test flaky (and failing on
headless).

Since this test is about the effect of changes in touch-action and not
in DOM, this CL removes this assumption by replacing the DOM change
with a touch-action change.

Additionally, this CL cleans up the test a bit, and adds attribute to
prevent text-selection and drag-and-drop in case these affect the
outcome of finger-drag in some platform.

Bug: 368980630
Change-Id: Ic0b4c49cd8409f4722cddf2f6b8424c501d6ec0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6012919
Reviewed-by: Kevin Ellis <keverschromium.org>
Commit-Queue: Mustaq Ahmed <mustaqchromium.org>
Cr-Commit-Position: refs/heads/main{#1382965}

--

wpt-commits: 881deaf0033cb15b7382a6911ca7d1d8460537d0
wpt-pr: 49172

UltraBlame original commit: a6793c4d17c61cad4a74e9295be3d40640381b31