Bug 1567058: part 2) Add tests for when the parent global has CSP and the target global hasn't. r=freddyb
The behavior is currently insufficently specified, see [1] and other
comments of that issue.

The new tests cover the cases of [2] where only the parent global has a
CSP. With additional tests for the <area> element.

For Chrome and Firefox all tests pass. For WebKit some fail because of
[3].

The set of tested APIs which accept `javascript:` URLs may not be
exhaustive. To my knowledge there is no explicit list of all APIs which
accept `javascript:` URLs.

Tests for when the target global has CSP and the source global has not
will be added in separate patches.

[1] https://github.com/whatwg/html/issues/4651#issuecomment-2355950176
[2] https://github.com/whatwg/html/issues/4651#issuecomment-495050351
[3] https://bugs.webkit.org/show_bug.cgi?id=13543

Differential Revision: https://phabricator.services.mozilla.com/D227223

UltraBlame original commit: 30e5049bce4199ff021bfdec23156f2c7c7b1fda