Bug 1692873 - Create smaller video elements for playback in simulcast tests. r=jib
For debugging purposes, when the test fails in CI and creates a screenshot.
gDM sources tend to be large, and for the sake of this test they should be
encoded large too. This shrinks the dimensions of the rendered video tracks
instead.

Differential Revision: https://phabricator.services.mozilla.com/D227492

UltraBlame original commit: 1b2720a24c441f1622ca58a16a1aedd1cc5ebc3a