Bug 1908330 - part 1 - Hide the navbar when entering search from browser r=android-reviewers,skhan
This applies to when we navigate to the homescreen which initializes it's navbar and then
immediately opens the search dialog.
Previously we would hide the entire bottom toolbar (addressbar + navbar) which could lead
to situations in which there is no toolbar shown on the screen until the one from search
is visible.

Differential Revision: https://phabricator.services.mozilla.com/D228179

UltraBlame original commit: 0a3a250e137507fc1d18b48d9092df1856a319ad