Bug 1908330 - part 5 - Remove BottomToolbarContainerIntegration from HomeFragment r=android-reviewers,harrisono
The `BottomToolbarContainerIntegration` through it's internal `ToolbarBehaviorController`
controls whether the toolbar should be dynamic or not depending on the state of the
current browser tab.
Since on the home screen there in no active browser tab and the toolbar from the home screen is
always fixed in place and not dynamic there is no need for `BottomToolbarContainerIntegration`.

Differential Revision: https://phabricator.services.mozilla.com/D228183

UltraBlame original commit: cf3a87ce6d2fe0477a8c66bad693f004b9f9081f