Bug 1932416 - Avoid shader miscompile due to if statement on Adreno devices. r=gfx-reviewers,nical
The brush_image shader handles normalized input UV coordinates by
first unnormalizing them if the flag is set, allowing the same logic
to be used subsequently regardless of whether the flag is
set. Unfortunately this appears to cause a miscompilation in the
REPETITION variant of the shader on some Adreno devices. Removing the
additional branch by replacing the if statement with a mix() avoids
the issue.

Differential Revision: https://phabricator.services.mozilla.com/D229796

UltraBlame original commit: 0d36e9e1ce0a0ef94a974839ad3a9296d013aaff