Bug 1917000 - Add permission message for "userScripts" permission r=zombie,fluent-reviewers,android-reviewers,zmckenney
This adds a permission message for the "userScripts" permission.

This also marks "userScripts" as an optional permission as a preparation
for implementing it as an optional-only permission. The implementation
and tests still expects the permission to have been granted at install
time; the next patches will update the tests and implementations to
complete the transition to an optional-only permission.

This feature is still behind a pref, off by default.

Differential Revision: https://phabricator.services.mozilla.com/D229711

UltraBlame original commit: bc4881030055df20d20bfe66fbd662e10884ce78