Bug 1925128 - Convert ParseRecordObject to a NativeObject r=mgaudet
This reduces the load per minor GC when parsing large JSON structures, which
requires a corresponding large (temporary) ParseRecordObject structure.

InternalizeJSONProperty() can be passed a null parseRecord pointer if, for
example, a reviver function modifies a value and then Steps 2a(iii)(1) or
2c(ii)(1) cannot find the id they are looking for. In the previous version of
the code, this was the same as a ParseRecordObject with undefined `value`
member.

The ParseRecordObject entries are stored as enumerated properties to avoid
having to allocate an extra JSObject.

Differential Revision: https://phabricator.services.mozilla.com/D228901

UltraBlame original commit: 44b15e68383a9e5ef48293cc91ae1df14e55096b