Bug 1922078 - Retrieve the functional parameter from element when collecting rules. r=view-transitions-reviewers,emilio
In order to match the selectors for the view-transition pseudo element,
we have to make sure we are using the correct `PseudoElement` (which includes
the functional parameter) when collecting the rules.

Note that this function is also used when matching the pseudo element in
the next patch.

Differential Revision: https://phabricator.services.mozilla.com/D230024

UltraBlame original commit: c288dcbae8bf00183a4a426ce7c5b3dfa03543ab