Bug 1930251 - Print a warning if we reach the mark stack limit while processing the test mark queue r=sfink
We don't need this to be an assertion failure if that's going to trip up
fuzzers. We can just print a warning instead. This is unlikely to happen
outside of testing and the code isn't present in release builds.

Differential Revision: https://phabricator.services.mozilla.com/D230993

UltraBlame original commit: a54c6fa20b54d0d19ab7d057b01eff8f1a4d19f7