Bug 1933557 part 5 - Optimize Map/Set constructor calls with an iterable argument in the JITs. r=iain
It's pretty common to call the constructor with a `null` or `undefined` argument and
with this patch we can stay in JIT code in this case.

If the argument is not `null` or `undefined` we can still allocate the object in
JIT code but we then call into C++ to do the initialization.

This could be optimized more later by emitting CacheIR guards to replace the
`IsOptimizableInitForMapOrSet` call.

Differential Revision: https://phabricator.services.mozilla.com/D230975

UltraBlame original commit: dfe8a8d407c463f290d3292d39916bc90011baee