Bug 1933756: Keep a strong reference to the taskbar icon controller as long as the window is alive. r=jhirsch
A window's icon controller is only held weakly by the widget code and once it is destroyed it
appears to completely remove the taskbar entry for that window. This maintains a strong reference
the the icon controller once created until the window that owns it is destroyed.

Differential Revision: https://phabricator.services.mozilla.com/D231089

UltraBlame original commit: e1b66b52b931e6cf43178434f0e9eee4a027e835