Bug 1924782 - Use --full option to show all available mach try perf tasks. r=perftest-reviewers,dom-storage-reviewers,janv,kshampur
This patch enables the use of --full to show all the available tasks instead of the categories. The `--show-all` still exists but it's destination is changed to `full` so that we can provide backwards compatibility with this new method. At the same time, all docs are updated to mention --full instead of, or alongside, --show-all.

Differential Revision: https://phabricator.services.mozilla.com/D230998

UltraBlame original commit: b74aa2545ad376cf98d69d4663f6e96ac516afbc