Bug 1935547 part 2: Introduce a "ComputeStretchContentBoxSize" helper to use in places where we're resolving something like the 'stretch' keyword. r=layout-reviewers,emilio
This patch doesn't change behavior[1]; it's just a refactor.

[1] Minor caveat: the "availBased" variables in this patch are now floored-to-0
whereas previously they were not.  In cases where these variables weren't
getting seprately-clamped, this is a bit of a behavior-change, but it's a step
towards correctness to floor this size-measurement to be nonnegative.

Differential Revision: https://phabricator.services.mozilla.com/D230228

UltraBlame original commit: ef053c45c33b2ad7d8dcf06cd4204499bd819036