Bug 1932772 - part9 : make it possible to fallback to WMF HW decoder in GPU process if ffmpeg fails to init a HW decoder. r=stransky,chunmin
Currently, in the GPU process, we attempt to create an FFmpeg decoder
prior to a WMF decoder. If the FFmpeg decoder fails to initialize its
HW decoder, we should fall back to the WMF decoder for HW decoding
instead.

We aim to avoid silently defaulting to the SW FFmpeg decoder in the GPU
process. To address this, this patch modifies the initialization
process, ensuring the HW decoder is initialized earlier when creating
an FFmpeg video decoder. This allows us to determine upfront whether HW
decoding via the FFmpeg decoder will succeed.

Differential Revision: https://phabricator.services.mozilla.com/D230343

UltraBlame original commit: 0630d8b0734cae9734f7a312f271ece348097db9