Bug 1936641, part 3 - Cycle collect SessionStoreChild::mManager in the child process. r=nika,farre,core-sessionstore-reviewers
Since bug 1875528, an IPDL actor maintains a strong reference to its manager.
SessionStoreChild is cycle collected. In the child process, its manager is BrowserChild,
which is also cycle collected. We need to tell the cycle collector about this edge to
ensure that the SessionStoreChild and its BrowserChild are collected at the same time.

This static cast looks questionable, given that the manager of PSessionStore can be
either PBrowser or PInProcess, but based on SessionStoreChild::GetOrCreate(), we
can tell the class of our manager by which process we're in.

Differential Revision: https://phabricator.services.mozilla.com/D231852

UltraBlame original commit: 7785305be9122b8e07cc585818c0e2c27b2a1959