Bug 1934118 - Tweak the returned value of GetSnapPointForDestination. r=dlrobertson
Since bug 1922904, GetSnapPointForDestination had returned an ineligible
snap point.

For exmaple, given that there's a scroll container only wants to snap
horizontally and there are 3 snap target points, (0, 0), (100, 0) and
(200,0).  Then if the user tries to scroll vertically to (0, 100), the
function returned (0, 0) as a valid snap position for the vertical scroll
operation. With this change the function now returns (0, 100) so that
the user can scroll vertically to the correct position.

Differential Revision: https://phabricator.services.mozilla.com/D231058

UltraBlame original commit: a3870f21337bd037347fe532c06864d0396652f8