Bug 1907126 - Split `--test-process-updates` functionality r=nrishel,nalexander
`--test-process-updates` currently basically does two totally different things: 1) It facilitates testing applying updates via the application startup process by preventing the application from actually starting if the update isn't applied or when it is re-launched as the callback application, and 2) If `ShouldNotProcessUpdates` finds a reason not to process updates, it does nearly the opposite: it sets the env variable to represent the reason we didn't process updates and lets the application continue to launch in order for a background task to consume that value.

This did not previously cause problems because we never did any testing that resulted in finding a reason not to process updates at startup that didn't involve also involve starting a background task. But the test later in this stack needs to do exactly that, motivating this change.

So this patch will split out functionality (2) into `--test-should-not-process-updates` so that the `--test-process-updates` behavior can consistently be: only process updates, do not launch the browser.

Differential Revision: https://phabricator.services.mozilla.com/D230013

UltraBlame original commit: 2c26ce3f8dd3c19876d8e25481ea853167ae42ed