Bug 1936004 Only clip once CSP violation sample for Trusted Types. r=smaug
Currently, ShouldSinkTypeMismatchViolationBeBlockedByCSP:
- Calculate trimmedSample (clipping to the first 40 characters).
- Set sample to the concatenation of sink name, "|", trimmedSample.
- Pass the sample to CSPViolationData's constructor, which clips again
  to the number of characters specified by the
  `security.csp.reporting.script-sample.max-length` preference, possibly
  appending an ellipsis and avoiding cut in the middle of a surrogate
  pair.

This patch removes the second clipping in CSPViolationData's constructor
as well as the extra ellipsis, so that our behavior is aligned with the
specifications and new WPT tests for the `eval` sink pass. For the
`Function` sink, follow-up work will be done in bug 1919729.

This patch also makes ShouldSinkTypeMismatchViolationBeBlockedByCSP
rely on the same clipping code as CSPViolationData, so that the
clipping length can be controlled by the corresponding preference and
we don't cut at a surrogate pair.

There are open questions about our mismatch with the CSP specification
that are not specific to Trusted Types, see bug 1935996 for that.
Corresponding TODOs are added in the code.

Differential Revision: https://phabricator.services.mozilla.com/D231912

UltraBlame original commit: 3491de781ce27c27f82eac63f2e089e457040065