Bug 1935714 - Part 1: Improve handling of large nsStringStream streams at the API level, r=mccr8,webdriver-reviewers,necko-reviewers,valentin,extension-reviewers,devtools-reviewers,robwu,ochameau
This changes how the `nsIStringInputStream` interface is initialized in
C++ code to make it support larger stream sizes better. This involved
moving away from using nsCString in some cases, as it has a maximum
string length of approximately 2GiB.

This also required changing the signature of one of the methods making
it no longer JS-compatible. As the `setData` method was awkward to use
from JS anyway, the method was split into two (one for C++ callers, and
one for JS callers), and renamed. This required changes to the various
JS callers.

Differential Revision: https://phabricator.services.mozilla.com/D231982

UltraBlame original commit: adc7412eeab1a4fcb8f10211a268b799c22350a3