Bug 1913421: Submit ping with at least 24 hour delay since last submission or on pref changes. r=mossop,geckoview-reviewers,tjr,tthibaud
We submit the ping with at least 24 hours delay in between to not spam with pings. When we receive a new captcha event,
we'll check the last submission time and submit the ping if it's been more than 24 hours since the last submission.

Differential Revision: https://phabricator.services.mozilla.com/D226023

UltraBlame original commit: da2a230542bdc71e7a2bf873a86f86d610827abc