Bug 1658072 - Preamble: refactor IsProcessDead. r=ipc-reviewers,nika
This patch rearranges the abstraction around `waitpid`/`waitid`.  There
is now a function `WaitForProcess` that returns more detailed information
about the process status, and then a wrapper `IsProcessDead` which is
private to the `ProcessWatcher` which simplifies the results and logs
information to stderr.  In the future, the extended information could be
exposed to observers within Gecko in some way.

Differential Revision: https://phabricator.services.mozilla.com/D230468

UltraBlame original commit: a4388793480850ce0a1e65c35ddd964bac08dc16