Bug 1937570 part 3 - Add Baseline CacheIR stub for JSOp::Lambda. r=iain
This is similar to the `NewObject` and `NewArray` IC stubs. We allocate a new
function object in JIT code and copy the shape and most slots from the canonical
function.

`js::Lambda` is one of our hottest `VMFunction`s on Speedometer 3 and this gets
rid of most of these calls. On Baseline-only micro-benchmarks this is almost 3x faster.

Differential Revision: https://phabricator.services.mozilla.com/D232348

UltraBlame original commit: 1dbbfc3b5ee2ef9ea5f9d0b9769e42f1817bdf97