Bug 1935924 - part 1: Implement `EditorLineBreak` class to treat both `<br>` and a preformatted linefeed with one data structure r=m_kato
`HTMLEditor` and its helper classes need to handle both a `<br>` and a
preformatted linefeed in the same way.  However, the former is an independent
node but the latter is a char in a `Text` node.  Therefore, it makes the
implementation simpler for making each `<br>` element handler treat a
preformatted linefeed in same path if there is a class to store the both
types of a line break.

Differential Revision: https://phabricator.services.mozilla.com/D231654

UltraBlame original commit: f55b5f143f8bc9450c3152fe7ccc8bd3f750af27