Bug 1935451 - perform seeking on the newly created state machine if there is an unfinished seeking on previous state machine. r=media-playback-reviewers,jolin
For ExternalEngineStateMachine, if the seeking happens before the engine
initialization, the seeking would not be executed immediately and be
postponed. If we switch to another state machine afterward, in
`SetStatusUpdateForNewlyCreatedStateMachineIfNeeded()` we will perform
the seeking again.

This patch makes decision to perform a pending seek by checking
`mLogicallySeeking`, not `mLogicalPosition`. That allows a pending
seeking to 0 second being able to run correctly.

Differential Revision: https://phabricator.services.mozilla.com/D232433

UltraBlame original commit: 9342060fcf1bfe16c15262bd43728092ac31fc24