Bug 1928120 - Defer VP8 decoder configuration when image size is unspecified r=media-playback-reviewers,alwu
This patch modifies the `VPXChangeMonitor::CanBeInstantiated()` to
postpone creating and initializing the VP8 decoder until the `decode()`
step if the image size is unknown during the `configure` step.

Currently, if `FFmpegVideoDecoder` for VP8 (via libvpx) does not receive
an image size beforehand, it will fail in `Init`, causing `VideoDecoder`
to return an error.

To prevent this situation, the patch updates
`VPXChangeMonitor::CanBeInstantiated` to check whether the image size is
known in advance. If it is not, decoder creation and initialization is
deferred until `Decode`, where the image size can be retrieved from the
input data.

This change ensures that the VP8 decoder can be successfully
`configure`d even when the image size is initially omitted.

Differential Revision: https://phabricator.services.mozilla.com/D232928

UltraBlame original commit: 759adf3f99df053708224d0f186dceecadd31403