Bug 1938588 - Fixing external object support, r=markh
Added code to check if a pointer type is for an external type and if so,
use the correct namespace.

It doesn't feel great to me to iterate over all external types for this.
Also the `crate_name_to_namespace` function seems hacky.  I hope to fix
this soon, maybe once the UniFFI bindgen pipline changes are merged and
vendored in.

Refactored some of the test crates to use proc-macros, which made the
testing easier.

Differential Revision: https://phabricator.services.mozilla.com/D232770

UltraBlame original commit: dabc5c0d3cdeb0e2a694f093d6dc00648d30b748