Bug 1933428 - Part 3: Recompute partitionKey for top-level documents if required. r=bvandersloot,smaug
The principal of a document might be changed during loading in certain
situation, e.g. a sandboxed document. In this case, we need to recompute
the partitionKey of the document to reflect the change on the principal.

Differential Revision: https://phabricator.services.mozilla.com/D230819

UltraBlame original commit: 9878302a4cb49d9f363b27b9986b2be2a58dbe68