Bug 1939342 part 2 - Support sparse elements in TryFastDeleteElementsForNewLength. r=iain
This is a lot faster for the case reported in the bug where the array has a few very large
sparse indexes, because we no longer iterate over the whole range.

This also lets us use the (very) fast path for dense elements if the object has a mix of
dense and sparse elements.

When we add a special representation for sparse elements it should be pretty easy to convert
this code to work on a hash-set instead.

Differential Revision: https://phabricator.services.mozilla.com/D233334

UltraBlame original commit: 738e5cf4b14f29bc86fb89da2820bd2a5a28ce6e