Bug 1938128 - test262-export: improve handling of unsupported features; r=dminor
And exclude two more early-stage proposals.

This reorganizes the code a bit so that the computation of the front matter is separated from inserting it into the exported source; this also makes it easier to skip tests based on the final feature flags rather than only on the ones extracted from the reftest skip-if line.

Differential Revision: https://phabricator.services.mozilla.com/D232525

UltraBlame original commit: fefde0c10da0cc27e55fbe88696e95090dc70706