Bug 1938490 - Do not enforce a single root group for documents. r=morgan
I believe this was needed in previous versions of VoiceOver, but now things seem to work.
We still need the root group for aria roles on the document, but that should be it.

This greatly accelerates our VoiceOver performance.

Differential Revision: https://phabricator.services.mozilla.com/D233759

UltraBlame original commit: c20079aa4d66d5e108ef5b5a3002730390a2891f