Bug 1940784 - Part 2: Lock at most once per buffer allocator GC operation r=sfink
This adds MaybeLock which is initially unlocked and only locked the first time
it is required.

This means that for some operations we may not lock at all. For example at the
start of minor collection we don't need to lock if we have already merged data
from previous collections back to the main thread. And if we do it's only once,
not once per zone.

Differential Revision: https://phabricator.services.mozilla.com/D233685

UltraBlame original commit: 3fd65f9a5866dc84f8ef2c027b9fc1a2fd4daf42