Bug 1882218 - Add embedder API for JSObject with non-GC pointer stashed in it. r=jandem
Adds new convenience API for usage by embedders, creating a JSObject
with a C++ pointer stashed in it, with the possibility to call a
callback to free the pointer when the JSObject is finalized.

Closely based on a design by Marco Trevisan.

Co-authored-by: Marco Trevisan (Treviño) <mail3v1n0.net>

Differential Revision: https://phabricator.services.mozilla.com/D231461

UltraBlame original commit: bac619a7d667291a7f6e9496f3406e8ed15dec24