Bug 1938340 - Improve early error handling and logging. r=gsvelto,fluent-reviewers
To better expose early errors, this rearranges the code to set the log
file as early as possible, and updates the file when we get more
information from parsing the extra file. Otherwise, some of the earlier
error paths might only be logging to stderr, which is difficult for
users access and provide to us in a bug.

This also adds a textbox with the full error source trace to the error
dialog. The trade-off here is that the first line will be localized
(when possible), however the rest of the error information will not be
(as the strings come from external code). This is fine, as this
information is mainly useful in bug reports (which is also why I made it
a textbox, for easy text selection).

Differential Revision: https://phabricator.services.mozilla.com/D233429

UltraBlame original commit: 0874fcba521fcf935ace34a3507f284de9bc4d79