Bug 1904906 - Write custom drag Modifier r=android-reviewers,matt-tighe
The SwipeToDismissBox is largely the same implementation. The main difference is moving away from `Modifier.anchoredDraggable` to a custom `Modifier.anchoredHorizontalDraggable`. This was inspired by a google issue tracker bug. This is feature flagged with the intent of gathering feedback while this bakes in Nightly and fine tuning the gesture based on UX and user feedback before letting it ride the trains.

Assuming this is stable and lands in Release, `SwipeToDismissBox` would be deleted and `SwipeToDismissBox2` would be renamed.

See Bugzilla for the full history of this bug.

https://issuetracker.google.com/issues/363198504

Differential Revision: https://phabricator.services.mozilla.com/D232774

UltraBlame original commit: 01f9fed7f6859f433503f7d14829246ed91516d8