Bug 1934715 - Add a filelock to `CommandSiteManager.ensure` to fix a race condition in `./mach lint` r=ahal,firefox-build-system-reviewers,glandium
During linting, two different commands were being invoked for the
`common` site, but the `common` site did not exist yet. One started
creating the site, the other slightly later, but noticed it wasn't
complete so tried to delete it to create it fresh. This cannot work.

To fix this, we use a file lock to ensure that only the first process to
reach the `ensure` point (which is the branch that may or may not create
it) has the ability to create it without interruption.

Differential Revision: https://phabricator.services.mozilla.com/D230875

UltraBlame original commit: 7af27f3eab4074f97e4f4945e19ede68751a6e16