Bug 1934827, part 2 - Text Fragments: Implement range content cache. r=farre
This is a prerequisite commit and introduces a data structure that is used in the following patches.

This patch introduces a helper function that converts the content of a `nsRange` into a fold case string.
Furthermore, it introduces a hashmap-based cache, which keeps the fold case representations accessible by `nsRange` pointer.

Differential Revision: https://phabricator.services.mozilla.com/D233387

UltraBlame original commit: 21b2f024f8898d3b072d31b52965a37b85ce29b3